refactor: Consolidate Drizzle migrations, fix date serialization, and update dependencies
- Replaced multiple Drizzle migration files with a single consolidated snapshot
- Fixed Date-to-string serialization in auth and email verification queries
- Updated package dependencies and Docker compose configuration
- Added /data/ to .gitignore for local data directory

+++ lib/data-access/orders.ts
@@ -127,7 +127,7 @@ export async function getOrdersByUserId(userId: number): Promise<OrderSummary[]>
   const counts = await db
     .select({
       orderId: orderItems.orderId,
-      count: sql<number>`COUNT(*)::int`,
+      count: sql<number>`COUNT(*)`,
     })
     .from(orderItems)
     .where(inArray(orderItems.orderId, orderIds))
